\name{clean}
\alias{clean}
\title{
	Extract only converged replications in the result objects
}
\description{
	Extract only the replications that are converegent in all supplied result objects (\code{\linkS4class{SimResult}})
}
\usage{
clean(...)
}
\arguments{
  \item{\dots}{
	The target result objects (\code{\linkS4class{SimResult}})
}
}
\value{
	The cleaned result objects
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
