\name{defaultStartingValues}
\alias{defaultStartingValues}
\title{
	Make ad hoc starting values
}
\description{
	Make ad hoc starting values to be analyzed by the \code{OpenMx} package
}
\usage{
defaultStartingValues(object)
}
\arguments{
  \item{object}{
	The \code{\linkS4class{SimParam}} object containing the analysis model
}
}
\value{
	The \code{\linkS4class{SimRSet}} object containing starting values
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
