\name{drawParameters}
\alias{drawParameters}
\title{
	Create parameter sets (with or without model misspecification) from the data object
}
\description{
	Create parameter sets (with or without model misspecification) from the data object, \code{\linkS4class{SimData}}
}
\usage{
drawParameters(object)
}
\arguments{
  \item{object}{
	The data object, \code{\linkS4class{SimData}}, used to draw a parameter set
}
}
\value{
	The list of parameters with model misspecification (\code{misspec}) and without model misspecification (\code{real})
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
