\name{find2Dhist}
\alias{find2Dhist}
\title{
	Fit the 2D Kernel Density Estimate
}
\description{
	Fit the 2D Kernel Density Estimate to a pair of variables
}
\usage{
find2Dhist(vec1, vec2) 
}
\arguments{
  \item{vec1}{
	Variable 1
}
  \item{vec2}{
	Variable 2
}
}
\value{
	The 2D Kernel Density Estimate based on each pair of values in \code{vec1} and \code{vec2}
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
