\name{Nullclass}
\docType{class}
\alias{NullDataFrame-class}
\alias{NullVector-class}
\alias{NullMatrix-class}
\alias{NullSimMatrix-class}
\alias{NullSymMatrix-class}
\alias{NullSimVector-class}
\alias{NullSimSet-class}
\alias{NullSimEqualCon-class}
\alias{NullSimREqualCon-class}
\alias{NullRSet-class}
\alias{NullSimMisspec-class}
\alias{NullSimDataDist-class}
\alias{NullSimMissing-class}
\alias{NullSimFunction-class}
\title{Null Objects}
\description{
List of all null objects. 
}
\section{Distributions}{
	Here is the list of all null objects and the link to their original classes.
	\itemize{
		\item \code{NullDataFrame} The null class from the \code{data.frame} class
		\item \code{NullVector} The null class from the \code{vector} class
		\item \code{NullMatrix} The null class from the \code{matrix} class
		\item \code{NullSimMatrix} The null class from the \code{\linkS4class{SimMatrix}} class
		\item \code{NullSymMatrix} The null class from the \code{\linkS4class{SymMatrix}} class
		\item \code{NullSimVector} The null class from the \code{\linkS4class{SimVector}} class
		\item \code{NullSimSet} The null class from the \code{\linkS4class{SimSet}} class
		\item \code{NullSimEqualCon} The null class from the \code{\linkS4class{SimEqualCon}} class
		\item \code{NullSimREqualCon} The null class from the \code{\linkS4class{SimREqualCon}} class
		\item \code{NullRSet} The null class from the \code{\linkS4class{VirtualRSet}} class
		\item \code{NullSimMisspec} The null class from the \code{\linkS4class{SimMisspec}} class
		\item \code{NullSimDataDist} The null class from the \code{\linkS4class{SimDataDist}} class
		\item \code{NullSimMissing} The null class from the \code{\linkS4class{SimMissing}} class
		\item \code{NullSimFunction} The null class from the \code{\linkS4class{SimFunction}} class
	}
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
\keyword{classes}
