\name{revText}
\alias{revText}
\title{
	Reverse the proportion value by subtracting it from 1
}
\description{
	Reverse the proportion value by subtracting it from 1. This function can reverse a value reported in text, such as from "> .98" to "< .02"
}
\usage{
revText(val)
}
\arguments{
  \item{val}{
	The value to be reversed
}
}
\value{
	The reversed value or text
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# This is a private function.

# revText(.96)
# revText("> .60")
}
