\name{twoTailedPValue}
\alias{twoTailedPValue}
\title{
Find two-tailed \emph{p} value from one-tailed \emph{p} value
}
\description{
Find two-tailed \emph{p} value from one-tailed \emph{p} value
}
\usage{
twoTailedPValue(vec)
}
\arguments{
  \item{vec}{
	A vector of one-tailed \emph{p} value.
}
}
\value{
	A vector of two-tailed \emph{p} value.
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
